--- app/javascript/application.js
@@ -1,7 +1,0 @@
-// Entry point for the build script in your package.json
-import { Turbo } from "@hotwired/turbo-rails";
-Turbo.session.drive = false;
-import "./controllers";
-
-
-import '../place_autocomplete.js';
