+++ app/views/public/posts/_form.html.erb
@@ -12,14 +12,17 @@
             <% end %>
           </ul>
         <% end %>
-        <%= form_with model: post do |f| %>
+        <%= form_with model: post, local: true do |f| %>
           <%= f.file_field :image, accept: "image/*", class: "w-100", autocomplete: "off" %>
           <div class="input-group mb-3 mt-4">
             <%= f.text_field :title, id: "Title", class: "form-control", placeholder: "タイトル", autocomplete: "off" %>
           </div>
+          <div>
           <%= f.text_area :body, id: "Body", class: "form-control", rows: "3", placeholder: "本文", autocomplete: "off" %>
-
+          </div>
+          <div>
           <%= f.text_field :address, id: "Address", class: "form-control", placeholder: "撮影地" %>
+          </div>
           <hr>
           <%= f.submit '投稿',class: "btn btn-lg btn-block btn-info mt-4" %>
         <% end %>
