+++ app/views/public/posts/new.html.erb
@@ -2,5 +2,3 @@
 <div class="container">
   <%= render "form", post: @post %>
 </div>
-
-<script src="https://maps.googleapis.com/maps/api/js?key=<%= ENV["GOOGLE_API_KEY"] %>&libraries=places&callback=initialize"></script>
